+++ frigate/comms/dispatcher.py
@@ -7,7 +7,7 @@ from typing import Any, Callable
 from frigate.config import FrigateConfig
 from frigate.ptz import OnvifCommandEnum, OnvifController
 from frigate.types import CameraMetricsTypes, FeatureMetricsTypes
-from frigate.util import restart_frigate
+from frigate.util.services import restart_frigate
 
 logger = logging.getLogger(__name__)
