+++ frigate/ffmpeg_presets.py
@@ -6,7 +6,7 @@ from enum import Enum
 from typing import Any
 
 from frigate.const import BTBN_PATH
-from frigate.util import vainfo_hwaccel
+from frigate.util.services import vainfo_hwaccel
 from frigate.version import VERSION
 
 logger = logging.getLogger(__name__)
