+++ frigate/track/norfair_tracker.py
@@ -7,7 +7,7 @@ from norfair.drawing.drawer import Drawer
 
 from frigate.config import DetectConfig
 from frigate.track import ObjectTracker
-from frigate.util import intersection_over_union
+from frigate.util.image import intersection_over_union
 
 
 # Normalizes distance from estimate relative to object size
